Move NMS into the modeling code
